Finish Day 15 part two
Two different approaches here. Started with implicit graph
representation in "ExpandedCaveGraph" and moved to an explicit
representation in the original CaveGraph implementation. My bugs turned
out to be in a bad implementation of the 1-9 wrapping, but interestingly
my sparse_lowest_cost_between performed much worse than the original
direct implementation with a Vec