Optimize volume tracking by removing completely contained regions.
If a region is completely contained then we'll generate an intersection
which is exactly that region and cancel to zero. Instead of doing this
we can drop the region entirely. This drops the runtime on the problem
from ~20 to ~2s. There's a step early on which intentionally(?) covers
most of the existing regions.