Properly implement A*/Dijkstra w/ dist map "memoization"
If we re-encounter a previously expanded state it could be already
expanded, or on the frontier. If it's expanded then we've found it with
a longer path, and we can drop it. If it's on the frontier we _might_
have found a shorter path and we should update our priority queue
accordingly.

This drops runtimes from ~60s to 1s and 7s for Parts 1 and 2
respectively. This is because we're (re)exploring many fewer states.

(I suppose you can view this as a kind of dynamic programming)